fix config tests
